% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_mcmc_draws}
\alias{get_mcmc_draws}
\title{Extract the MCMC draws from a glam_mcmc object}
\usage{
get_mcmc_draws(x)
}
\arguments{
\item{x}{an object of class \code{glam_mcmc}, as produced by
\code{glam_mcmc$new()}.}
}
\description{
Raw MCMC draws are stored within an object of class
  \code{glam_mcmc} in a series of lists. This makes it easier to extend the
  MCMC as needed, but is not the most convenient format for exploring or
  plotting. This function extracts these draws into a convenient long-form
  data.frame for use in downstream functions.
}
