% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ind}
\alias{plot_ind}
\title{Plot a single simulated individual}
\usage{
plot_ind(ind)
}
\arguments{
\item{ind}{a single individual simulated using \code{sim_ind()} with
\code{return_full = TRUE}.}
}
\description{
Produces a plot showing the infections in a single individual,
  and which haplotypes were introduced in each infection. Requires that the
  individual was simulated with hidden states returned (\code{return_full =
  TRUE}).
}
