% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_ind}
\alias{sim_ind}
\title{Simulate a single individual}
\usage{
sim_ind(
  samp_time,
  haplo_freqs,
  lambda,
  theta,
  decay_rate,
  sens,
  ind_name = "ind1",
  n_inf = NULL,
  return_full = FALSE
)
}
\arguments{
\item{samp_time}{numeric vector of times at which observations were taken.
Must be at least two unique time points increasing in value.}

\item{haplo_freqs}{vector of population-level haplotype frequencies.}

\item{lambda}{force of infection.}

\item{theta}{COI intensity parameter.}

\item{decay_rate}{rate at which haplotypes clear.}

\item{sens}{sensitivity of sequencing. Assumed the same for all haplotypes.}

\item{ind_name}{the name given to this individual in the returned data.frame.}

\item{n_inf}{if \code{NULL} then number of infections is drawn from the
model, but this parameter can also be used to manually set this value.}

\item{return_full}{Boolean (Default \code{FALSE}). If \code{TRUE} then
return true values of hidden variables alongside observed data.}
}
\description{
Simulate a single individual from the generative model. Returns
  outputs as a list, including observed data as well as optionally hidden
  true states. If hidden states are returned then \code{plot_ind()} can be
  used to visualise this simulation.
}
\details{
Note that the method for simulating data is deliberately different
  from the way that it is represented in the inference step, although it is
  mathematically equivalent. This allows us to check that our different
  representations of the process are consistent.
}
