% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLAM.R
\docType{package}
\name{GLAM}
\alias{GLAM-package}
\alias{GLAM}
\title{GLAM}
\description{
Malaria infections often contain multiple haplotypes. When
sequenced at several points in time, these haplotypes may be present for a
period beforing cleaing as parasite densities decrease. We may also fail to
detect a haplotype at a given point in time due to imperfect sensitivity of
sequencing. GLAM takes these factors into account in a probabilistic model
and produces estimates of the time(s) at which each individual in a
longitudinal cohort became infected, for how long this infection lasted,
and various other parameters.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mrc-ide/GLAM}
}

}
\author{
\strong{Maintainer}: Bob Verity \email{r.verity@imperial.ac.uk}

Authors:
\itemize{
  \item Sophie Berube \email{sberube3@jhmi.edu}
}

}
