% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{glam_mcmc}
\alias{glam_mcmc}
\title{R6 class representing a GLAM MCMC object}
\description{
Create a new glam_mcmc object to run an MCMC.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{example_public_var}}{example.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-glam_mcmc-new}{\code{glam_mcmc$new()}}
\item \href{#method-glam_mcmc-params}{\code{glam_mcmc$params()}}
\item \href{#method-glam_mcmc-burn}{\code{glam_mcmc$burn()}}
\item \href{#method-glam_mcmc-print}{\code{glam_mcmc$print()}}
\item \href{#method-glam_mcmc-clone}{\code{glam_mcmc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-glam_mcmc-new"></a>}}
\if{latex}{\out{\hypertarget{method-glam_mcmc-new}{}}}
\subsection{Method \code{new()}}{
Load data into the glam_mcmc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glam_mcmc$new(df_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_data}}{a data.frame of data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-glam_mcmc-params"></a>}}
\if{latex}{\out{\hypertarget{method-glam_mcmc-params}{}}}
\subsection{Method \code{params()}}{
Define model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glam_mcmc$params(
  chains = 5,
  rungs = 1,
  lambda = NULL,
  theta = NULL,
  decay_rate = NULL,
  sens = NULL,
  n_infections = NULL,
  haplo_freqs = NULL,
  max_infections = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chains}}{the number of independent chains.}

\item{\code{rungs}}{TODO}

\item{\code{lambda}}{TODO}

\item{\code{theta}}{TODO}

\item{\code{decay_rate}}{TODO}

\item{\code{sens}}{TODO}

\item{\code{n_infections}}{TODO}

\item{\code{haplo_freqs}}{TODO}

\item{\code{max_infections}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-glam_mcmc-burn"></a>}}
\if{latex}{\out{\hypertarget{method-glam_mcmc-burn}{}}}
\subsection{Method \code{burn()}}{
Run burn-in MCMC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glam_mcmc$burn(iterations, target_acceptance = 0.44, silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{number of burn-in iterations}

\item{\code{target_acceptance}}{all Metropolis-Hastings proposals will be
adaptively tuned to aim for this acceptance rate.}

\item{\code{silent}}{if \code{TRUE} then suppress all console output. Default =
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-glam_mcmc-print"></a>}}
\if{latex}{\out{\hypertarget{method-glam_mcmc-print}{}}}
\subsection{Method \code{print()}}{
Print MCMC object summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glam_mcmc$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-glam_mcmc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-glam_mcmc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glam_mcmc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
